+++ hutool-cache/src/main/java/cn/hutool/cache/impl/CacheObj.java
@@ -44,9 +44,8 @@ public class CacheObj<K, V> implements Serializable{
 	 */
 	boolean isExpired() {
 		if(this.ttl > 0) {
-			final long expiredTime = this.lastAccess + this.ttl;
-			// expiredTime > 0 杜绝Long类型溢出变负数问题，当当前时间超过过期时间，表示过期
-			return expiredTime > 0 && expiredTime < System.currentTimeMillis();
+			// 此处不考虑时间回拨
+			return (System.currentTimeMillis() - this.lastAccess) > this.ttl;
 		}
 		return false;
 	}