+++ hutool-json/src/main/java/cn/hutool/json/JSONObject.java
@@ -261,7 +261,8 @@ public class JSONObject implements JSON, JSONGetter<String>, Map<String, Object>
 	}
 
 	/**
-	 * 设置转为字符串时的日期格式，默认为时间戳（null值）
+	 * 设置转为字符串时的日期格式，默认为时间戳（null值）<br>
+	 * 此方法设置的日期格式仅对转换为JSON字符串有效，对解析JSON为bean无效。
 	 *
 	 * @param format 格式，null表示使用时间戳
 	 * @return this
@@ -341,7 +342,7 @@ public class JSONObject implements JSON, JSONGetter<String>, Map<String, Object>
 	}
 
 	/**
-	 * PUT 键值对到JSONObject中，在忽略null模式下，如果值为<code>null</code>，将此键移除
+	 * PUT 键值对到JSONObject中，在忽略null模式下，如果值为{@code null}，将此键移除
 	 *
 	 * @param key   键
 	 * @param value 值对象. 可以是以下类型: Boolean, Double, Integer, JSONArray, JSONObject, Long, String, or the JSONNull.NULL.
@@ -356,7 +357,7 @@ public class JSONObject implements JSON, JSONGetter<String>, Map<String, Object>
 	}
 
 	/**
-	 * 设置键值对到JSONObject中，在忽略null模式下，如果值为<code>null</code>，将此键移除
+	 * 设置键值对到JSONObject中，在忽略null模式下，如果值为{@code null}，将此键移除
 	 *
 	 * @param key   键
 	 * @param value 值对象. 可以是以下类型: Boolean, Double, Integer, JSONArray, JSONObject, Long, String, or the JSONNull.NULL.
@@ -426,7 +427,7 @@ public class JSONObject implements JSON, JSONGetter<String>, Map<String, Object>
 	 * @param key   键
 	 * @param value 被积累的值
 	 * @return this.
-	 * @throws JSONException 如果给定键为<code>null</code>或者键对应的值存在且为非JSONArray
+	 * @throws JSONException 如果给定键为{@code null}或者键对应的值存在且为非JSONArray
 	 */
 	public JSONObject accumulate(String key, Object value) throws JSONException {
 		InternalJSONUtil.testValidity(value);
@@ -447,7 +448,7 @@ public class JSONObject implements JSON, JSONGetter<String>, Map<String, Object>
 	 * @param key   键
 	 * @param value 值
 	 * @return this.
-	 * @throws JSONException 如果给定键为<code>null</code>或者键对应的值存在且为非JSONArray
+	 * @throws JSONException 如果给定键为{@code null}或者键对应的值存在且为非JSONArray
 	 */
 	public JSONObject append(String key, Object value) throws JSONException {
 		InternalJSONUtil.testValidity(value);
@@ -545,7 +546,7 @@ public class JSONObject implements JSON, JSONGetter<String>, Map<String, Object>
 
 	/**
 	 * 返回JSON字符串<br>
-	 * 如果解析错误，返回<code>null</code>
+	 * 如果解析错误，返回{@code null}
 	 *
 	 * @return JSON字符串
 	 */