+++ hutool-dfa/src/main/java/cn/hutool/dfa/FoundWord.java
@@ -1,49 +1,61 @@
 package cn.hutool.dfa;
 
-/**
- * @author 肖海斌
- * <p>
- * 匹配到的敏感词，包含敏感词，text中匹配敏感词的内容，以及匹配内容在text中的下标，
- * 下标可以用来做敏感词的进一步处理，如果替换成**
- */
-public class FoundWord {
-	/**
-	 * 生效的敏感词
-	 */
-	private String word;
-	/**
-	 * 敏感词匹配到的内容
-	 */
-	private String foundWord;
-	/**
-	 * 匹配内容在待分析字符串中的开始位置
-	 */
-	private int startIndex;
-	/**
-	 * 匹配内容在待分析字符串中的结束位置
-	 */
-	private int endIndex;
+import cn.hutool.core.lang.DefaultSegment;
 
-	public FoundWord(String word, String foundWord, int start, int end) {
+/**
+ * <p>
+ * 匹配到的单词，包含单词，text中匹配单词的内容，以及匹配内容在text中的下标，
+ * 下标可以用来做单词的进一步处理，如果替换成**
+ *
+ * @author 肖海斌
+ */
+public class FoundWord extends DefaultSegment<Integer> {
+	/**
+	 * 生效的单词，即单词树中的词
+	 */
+	private final String word;
+	/**
+	 * 单词匹配到的内容，即文中的单词
+	 */
+	private final String foundWord;
+
+	/**
+	 * 构造
+	 *
+	 * @param word 生效的单词，即单词树中的词
+	 * @param foundWord 单词匹配到的内容，即文中的单词
+	 * @param startIndex 起始位置（包含）
+	 * @param endIndex 结束位置（包含）
+	 */
+	public FoundWord(String word, String foundWord, int startIndex, int endIndex) {
+		super(startIndex, endIndex);
 		this.word = word;
 		this.foundWord = foundWord;
-		this.startIndex = start;
-		this.endIndex = end;
 	}
 
+	/**
+	 * 获取生效的单词，即单词树中的词
+	 *
+	 * @return 生效的单词
+	 */
 	public String getWord() {
 		return word;
 	}
 
+	/**
+	 * 获取单词匹配到的内容，即文中的单词
+	 * @return 单词匹配到的内容
+	 */
 	public String getFoundWord() {
 		return foundWord;
 	}
 
-	public int getStartIndex() {
-		return startIndex;
-	}
-
-	public int getEndIndex() {
-		return endIndex;
+	/**
+	 * 默认的，只输出匹配到的关键字
+	 * @return 匹配到的关键字
+	 */
+	@Override
+	public String toString() {
+		return this.foundWord;
 	}
 }