+++ hutool-dfa/src/main/java/cn/hutool/dfa/WordTree.java
@@ -1,11 +1,17 @@
 package cn.hutool.dfa;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.core.lang.Filter;
 import cn.hutool.core.text.StrBuilder;
 import cn.hutool.core.util.StrUtil;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 /**
  * DFA（Deterministic Finite Automaton 确定有穷自动机）
@@ -26,7 +32,7 @@ public class WordTree extends HashMap<Character, WordTree> {
 	private static final long serialVersionUID = -4646423269465809276L;
 
 	/**
-	 * 敏感词字符末尾标识，用于标识单词末尾字符
+	 * 单词字符末尾标识，用于标识单词末尾字符
 	 */
 	private final Set<Character> endCharacterSet = new HashSet<>();
 	/**
@@ -62,26 +68,30 @@ public class WordTree extends HashMap<Character, WordTree> {
 	 * 增加一组单词
 	 *
 	 * @param words 单词集合
+	 * @return this
 	 */
-	public void addWords(Collection<String> words) {
+	public WordTree addWords(Collection<String> words) {
 		if (false == (words instanceof Set)) {
 			words = new HashSet<>(words);
 		}
 		for (String word : words) {
 			addWord(word);
 		}
+		return this;
 	}
 
 	/**
 	 * 增加一组单词
 	 *
 	 * @param words 单词数组
+	 *              @return this
 	 */
-	public void addWords(String... words) {
+	public WordTree addWords(String... words) {
 		HashSet<String> wordsSet = CollectionUtil.newHashSet(words);
 		for (String word : wordsSet) {
 			addWord(word);
 		}
+		return this;
 	}
 
 	/**
@@ -89,7 +99,7 @@ public class WordTree extends HashMap<Character, WordTree> {
 	 *
 	 * @param word 单词
 	 */
-	public void addWord(String word) {
+	public WordTree addWord(String word) {
 		final Filter<Character> charFilter = this.charFilter;
 		WordTree parent = null;
 		WordTree current = this;
@@ -112,8 +122,8 @@ public class WordTree extends HashMap<Character, WordTree> {
 		if (null != parent) {
 			parent.setEnd(currentChar);
 		}
+		return this;
 	}
-
 	//------------------------------------------------------------------------------- match
 
 	/**
@@ -126,7 +136,7 @@ public class WordTree extends HashMap<Character, WordTree> {
 		if (null == text) {
 			return false;
 		}
-		return null != match(text);
+		return null != matchWord(text);
 	}
 
 	/**
@@ -135,15 +145,24 @@ public class WordTree extends HashMap<Character, WordTree> {
 	 * @param text 被检查的文本
 	 * @return 匹配到的关键字
 	 */
-	public FoundWord match(String text) {
+	public String match(String text) {
+		final FoundWord foundWord = matchWord(text);
+		return null != foundWord ? foundWord.toString() : null;
+	}
+
+	/**
+	 * 获得第一个匹配的关键字
+	 *
+	 * @param text 被检查的文本
+	 * @return 匹配到的关键字
+	 * @since 5.5.3
+	 */
+	public FoundWord matchWord(String text) {
 		if (null == text) {
 			return null;
 		}
-		List<FoundWord> matchAll = matchAll(text, 1);
-		if (CollectionUtil.isNotEmpty(matchAll)) {
-			return matchAll.get(0);
-		}
-		return null;
+		final List<FoundWord> matchAll = matchAllWords(text, 1);
+		return CollUtil.get(matchAll, 0);
 	}
 
 	//------------------------------------------------------------------------------- match all
@@ -154,8 +173,19 @@ public class WordTree extends HashMap<Character, WordTree> {
 	 * @param text 被检查的文本
 	 * @return 匹配的词列表
 	 */
-	public List<FoundWord> matchAll(String text) {
+	public List<String> matchAll(String text) {
 		return matchAll(text, -1);
+	}
+
+	/**
+	 * 找出所有匹配的关键字
+	 *
+	 * @param text 被检查的文本
+	 * @return 匹配的词列表
+	 * @since 5.5.3
+	 */
+	public List<FoundWord> matchAllWords(String text) {
+		return matchAllWords(text, -1);
 	}
 
 	/**
@@ -165,8 +195,20 @@ public class WordTree extends HashMap<Character, WordTree> {
 	 * @param limit 限制匹配个数
 	 * @return 匹配的词列表
 	 */
-	public List<FoundWord> matchAll(String text, int limit) {
+	public List<String> matchAll(String text, int limit) {
 		return matchAll(text, limit, false, false);
+	}
+
+	/**
+	 * 找出所有匹配的关键字
+	 *
+	 * @param text  被检查的文本
+	 * @param limit 限制匹配个数
+	 * @return 匹配的词列表
+	 * @since 5.5.3
+	 */
+	public List<FoundWord> matchAllWords(String text, int limit) {
+		return matchAllWords(text, limit, false, false);
 	}
 
 	/**
@@ -180,7 +222,24 @@ public class WordTree extends HashMap<Character, WordTree> {
 	 * @param isGreedMatch   是否使用贪婪匹配（最长匹配）原则
 	 * @return 匹配的词列表
 	 */
-	public List<FoundWord> matchAll(String text, int limit, boolean isDensityMatch, boolean isGreedMatch) {
+	public List<String> matchAll(String text, int limit, boolean isDensityMatch, boolean isGreedMatch) {
+		final List<FoundWord> matchAllWords = matchAllWords(text, limit, isDensityMatch, isGreedMatch);
+		return CollUtil.map(matchAllWords, FoundWord::toString, true);
+	}
+
+	/**
+	 * 找出所有匹配的关键字<br>
+	 * 密集匹配原则：假如关键词有 ab,b，文本是abab，将匹配 [ab,b,ab]<br>
+	 * 贪婪匹配（最长匹配）原则：假如关键字a,ab，最长匹配将匹配[a, ab]
+	 *
+	 * @param text           被检查的文本
+	 * @param limit          限制匹配个数
+	 * @param isDensityMatch 是否使用密集匹配原则
+	 * @param isGreedMatch   是否使用贪婪匹配（最长匹配）原则
+	 * @return 匹配的词列表
+	 * @since 5.5.3
+	 */
+	public List<FoundWord> matchAllWords(String text, int limit, boolean isDensityMatch, boolean isGreedMatch) {
 		if (null == text) {
 			return null;
 		}
@@ -239,8 +298,6 @@ public class WordTree extends HashMap<Character, WordTree> {
 		}
 		return foundWords;
 	}
-
-
 	//--------------------------------------------------------------------------------------- Private method start
 
 	/**